+++ homeassistant/components/renson_healthbox3/config_flow.py
@@ -13,9 +13,7 @@ import voluptuous as vol
 
 from homeassistant import config_entries
 from homeassistant.const import CONF_API_KEY, CONF_HOST
-from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import FlowResult
-from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.aiohttp_client import async_create_clientsession
 
 from .const import DOMAIN, LOGGER
@@ -33,21 +31,6 @@ class ConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
 
     VERSION = 1
 
-    async def validate_input(
-        self, hass: HomeAssistant, data: dict[str, Any]
-    ) -> dict[str, Any]:
-        """Validate the user input allows us to connect."""
-
-        if CONF_API_KEY in data:
-            await self._test_credentials(
-                ipaddress=data[CONF_HOST],
-                apikey=data[CONF_API_KEY],
-            )
-        else:
-            await self._test_connectivity(ipaddress=data[CONF_HOST])
-
-        return {"title": "Renson"}
-
     async def async_step_user(
         self, user_input: dict[str, Any] | None = None
     ) -> FlowResult:
@@ -60,9 +43,16 @@ class ConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
         errors = {}
 
         try:
-            info = await self.validate_input(self.hass, user_input)
-        except CannotConnect:
-            errors["base"] = "cannot_connect"
+            client = Healthbox3(
+                host=user_input[CONF_HOST],
+                api_key=user_input[CONF_API_KEY],
+                session=async_create_clientsession(self.hass),
+            )
+
+            if CONF_API_KEY in user_input:
+                await client.async_enable_advanced_api_features()
+            else:
+                await client.async_validate_connectivity()
         except Healthbox3ApiClientAuthenticationError:
             errors[CONF_API_KEY] = "auth"
         except Healthbox3ApiClientCommunicationError:
@@ -78,25 +68,3 @@ class ConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
         return self.async_show_form(
             step_id="user", data_schema=STEP_USER_DATA_SCHEMA, errors=errors
         )
-
-    async def _test_credentials(self, ipaddress: str, apikey: str) -> None:
-        """Validate credentials."""
-        client = Healthbox3(
-            host=ipaddress,
-            api_key=apikey,
-            session=async_create_clientsession(self.hass),
-        )
-        await client.async_enable_advanced_api_features()
-
-    async def _test_connectivity(self, ipaddress: str) -> None:
-        """Validate connectivity."""
-        client = Healthbox3(
-            host=ipaddress,
-            api_key=None,
-            session=async_create_clientsession(self.hass),
-        )
-        await client.async_validate_connectivity()
-
-
-class CannotConnect(HomeAssistantError):
-    """Error to indicate we cannot connect."""
